@@ -17,7 +17,7 @@ import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.PostConstruct;
+import jakarta.annotation.PostConstruct;
 import java.time.Clock;
 import java.time.OffsetDateTime;
 import java.util.List;
@@ -40,21 +40,13 @@ public class JobService {
 
     private static final Logger LOG = LoggerFactory.getLogger(JobService.class);
 
-    @Autowired
     private ApplicationEventPublisher applicationEventPublisher;
-    @Autowired
     private JobRepository jobRepository;
-    @Autowired
     private JobMetaService jobMetaService;
-    @Autowired
     private ScheduledExecutorService executor;
-    @Autowired(required = false)
     private List<JobRunnable> jobRunnables = emptyList();
-    @Autowired
     private UuidProvider uuidProvider;
 
-
-    @Autowired
     private SystemInfo systemInfo;
 
     private Clock clock = Clock.systemDefaultZone();
@@ -62,20 +54,21 @@ public class JobService {
     public JobService() {
     }
 
+    @Autowired
     JobService(final JobRepository jobRepository,
                final JobMetaService jobMetaService,
-               final List<JobRunnable> jobRunnables,
+               @Autowired(required = false) final List<JobRunnable> jobRunnables,
                final ScheduledExecutorService executor,
                final ApplicationEventPublisher applicationEventPublisher,
-               final Clock clock,
+               @Autowired(required = false) final Clock clock,
                final SystemInfo systemInfo,
                final UuidProvider uuidProvider) {
         this.jobRepository = jobRepository;
         this.jobMetaService = jobMetaService;
-        this.jobRunnables = jobRunnables;
+        this.jobRunnables = jobRunnables != null ? jobRunnables : this.jobRunnables;
         this.executor = executor;
         this.applicationEventPublisher = applicationEventPublisher;
-        this.clock = clock;
+        this.clock = clock != null ? clock : this.clock;
         this.systemInfo = systemInfo;
         this.uuidProvider = uuidProvider;
     }
@@ -155,13 +148,13 @@ public class JobService {
      */
     private void clearRunLocks() {
         jobMetaService.runningJobs().forEach((RunningJob runningJob) -> {
-            final Optional<JobInfo> jobInfoOptional = jobRepository.findOne(runningJob.jobId);
+            final Optional<JobInfo> jobInfoOptional = jobRepository.findOne(runningJob.jobId());
             if (jobInfoOptional.isPresent() && jobInfoOptional.get().isStopped()) {
-                jobMetaService.releaseRunLock(runningJob.jobType);
-                LOG.error("Clear Lock of Job {}. Job stopped already.", runningJob.jobType);
+                jobMetaService.releaseRunLock(runningJob.jobType());
+                LOG.error("Clear Lock of Job {}. Job stopped already.", runningJob.jobType());
             } else if (!jobInfoOptional.isPresent()) {
-                jobMetaService.releaseRunLock(runningJob.jobType);
-                LOG.error("Clear Lock of Job {}. JobID does not exist", runningJob.jobType);
+                jobMetaService.releaseRunLock(runningJob.jobType());
+                LOG.error("Clear Lock of Job {}. JobID does not exist", runningJob.jobType());
             }
         });
     }
@@ -264,8 +257,8 @@ public class JobService {
 
     public void handleTooBigJobLogs() {
         jobMetaService.runningJobs().forEach((RunningJob runningJob) -> {
-            LOG.debug("Keeping job messages small for job id {}", runningJob.jobId);
-            jobRepository.keepJobMessagesWithinMaximumSize(runningJob.jobId);
+            LOG.debug("Keeping job messages small for job id {}", runningJob.jobId());
+            jobRepository.keepJobMessagesWithinMaximumSize(runningJob.jobId());
         });
     }
 }
