@@ -2,6 +2,6 @@ package de.otto.edison.status.scheduler;
 
 public interface Scheduler {
 
-    public void update();
+    void update();
 
 }
