@@ -2,7 +2,6 @@ package de.otto.edison.jobs.definition;
 
 
 import org.springframework.scheduling.support.CronExpression;
-import org.springframework.scheduling.support.CronSequenceGenerator;
 
 import java.time.Duration;
 import java.util.Optional;
@@ -71,7 +70,7 @@ public final class DefaultJobDefinition implements JobDefinition {
      * @param jobType     The type of the Job
      * @param jobName     A human readable name of the Job
      * @param description A human readable description of the Job.
-     * @param cron        The cron expression. Must conform to {@link CronSequenceGenerator}s cron expressions.
+     * @param cron        The cron expression. Must conform to {@link CronExpression}s cron expressions.
      * @param restarts    The number of restarts if the job failed because of errors or exceptions
      * @param maxAge      Optional maximum age of a job. When the job is not run for longer than this duration,
      *                    a warning is displayed on the status page
@@ -93,7 +92,7 @@ public final class DefaultJobDefinition implements JobDefinition {
      * @param jobType     The type of the Job
      * @param jobName     A human readable name of the Job
      * @param description A human readable description of the Job.
-     * @param cron        The cron expression. Must conform to {@link CronSequenceGenerator}s cron expressions.
+     * @param cron        The cron expression. Must conform to {@link CronExpression}s cron expressions.
      * @param restarts    The number of restarts if the job failed because of errors or exceptions
      * @param retries     Specifies how often a job trigger should retry to start the job if triggering fails for some reason.
      * @param retryDelay  The optional delay between retries.
@@ -183,7 +182,7 @@ public final class DefaultJobDefinition implements JobDefinition {
     /**
      * @param cron cron expression to validate.
      * @throws IllegalArgumentException if cron expression is invalid
-     * @see CronSequenceGenerator
+     * @see CronExpression
      */
     public static void validateCron(String cron) {
         //Internally, we use the org.springframework.scheduling.support.CronTrigger which has a different syntax than the official crontab syntax.
