@@ -73,7 +73,7 @@ class DynamoJobRepositoryTest {
     }
 
     private static DynamoDbClient getDynamoDbClient() {
-        String endpointUri = "http://" + dynamodb.getContainerIpAddress() + ":" +
+        String endpointUri = "http://" + dynamodb.getHost() + ":" +
                 dynamodb.getMappedPort(8000);
 
         return DynamoDbClient.builder()
