@@ -115,7 +115,7 @@ public class JobMetaRepositoryTest {
     }
 
     private static DynamoDbClient getDynamoDbClient() {
-        String endpointUri = "http://" + dynamodb.getContainerIpAddress() + ":" +
+        String endpointUri = "http://" + dynamodb.getHost() + ":" +
                 dynamodb.getMappedPort(8000);
         return DynamoDbClient.builder()
                 .endpointOverride(URI.create(endpointUri))
