@@ -5,6 +5,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.ActiveProfiles;
@@ -24,14 +25,14 @@ public class ExampleStatusSmokeTest {
     @Test
     public void shouldRenderMainPage() {
         final ResponseEntity<String> response = this.restTemplate.getForEntity("/", String.class);
-        assertThat(response.getStatusCodeValue()).isEqualTo(200);
+        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
         assertThat(response.getBody()).startsWith("<html");
     }
 
     @Test
     public void shouldHaveStatusEndpoint() {
         final ResponseEntity<String> response = this.restTemplate.getForEntity("/internal/status", String.class);
-        assertThat(response.getStatusCodeValue()).isEqualTo(200);
+        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
         assertThat(response.getHeaders().getContentType()).isEqualTo(MediaType.APPLICATION_JSON);
         assertThat(response.getBody()).startsWith("{");
     }
@@ -40,7 +41,7 @@ public class ExampleStatusSmokeTest {
     public void shouldHaveHealthCheck() {
         final ResponseEntity<String> response = this.restTemplate.getForEntity("/actuator/health", String.class);
         assertThat(response.getHeaders().getContentType()).isEqualTo(MediaType.APPLICATION_JSON);
-        assertThat(response.getStatusCodeValue()).isIn(200, 503);
+        assertThat(response.getStatusCode().value()).isIn(200, 503);
     }
 
 }
